+++ packages/ai/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Added MiniMax M3 to the MiniMax Token Plan catalog and made China Token Plan credentials unlock the Anthropic-compatible `minimax-cn` SKU. ([#1790](https://github.com/can1357/oh-my-pi/issues/1790))
+
 ## [15.8.2] - 2026-06-03
 
 ### Fixed